简化网络请求，添加上传下载示例

+++ tcraftlib/src/main/java/com/tincher/tcraftlib/network/download/DownloadListener.java
@@ -6,6 +6,6 @@ package com.tincher.tcraftlib.network.download;
 
 public interface DownloadListener {
 
-    void onProgress(float progress,long total);
+    void onProgress(int progress,long total);
 
 }
